[PM-20195] Migrate SendsService to network module
Moved `SendsService` and its related classes (`CreateFileSendResponse`, `CreateSendJsonResponse`, `InvalidJsonResponse`, `UpdateSendResponseJson`, `SendsServiceImpl`) from the `app` module to the `network` module. Updated all usages of them to use the new implementations in the network. Removed duplicate implementations and interfaces from the `app` module.
