+++ network/src/main/kotlin/com/bitwarden/network/service/SendsServiceImpl.kt
@@ -1,19 +1,19 @@
-package com.x8bit.bitwarden.data.vault.datasource.network.service
+package com.bitwarden.network.service
 
 import androidx.core.net.toUri
 import com.bitwarden.network.api.AzureApi
 import com.bitwarden.network.api.SendsApi
+import com.bitwarden.network.model.CreateFileSendResponse
 import com.bitwarden.network.model.CreateFileSendResponseJson
+import com.bitwarden.network.model.CreateSendJsonResponse
 import com.bitwarden.network.model.FileUploadType
 import com.bitwarden.network.model.SendJsonRequest
 import com.bitwarden.network.model.SyncResponseJson
+import com.bitwarden.network.model.UpdateSendResponseJson
 import com.bitwarden.network.model.toBitwardenError
 import com.bitwarden.network.util.NetworkErrorCode
 import com.bitwarden.network.util.parseErrorBodyOrNull
 import com.bitwarden.network.util.toResult
-import com.x8bit.bitwarden.data.vault.datasource.network.model.CreateFileSendResponse
-import com.x8bit.bitwarden.data.vault.datasource.network.model.CreateSendJsonResponse
-import com.x8bit.bitwarden.data.vault.datasource.network.model.UpdateSendResponseJson
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.MultipartBody
@@ -107,7 +107,7 @@ class SendsServiceImpl(
                                     contentType = "application/octet-stream".toMediaType(),
                                 ),
                                 name = "data",
-                                filename = send.file?.fileName,
+                                filename = send.file.fileName,
                             ),
                         )
                         .build(),
